# src/core/core/static/openapi3_0.yaml
openapi: 3.0.3
info:
  title: Taranis
  contact:
    name: Benjamin Akhras
    email: benjamin.akhras@ait.ac.at
  description: |-
    The Taranis API provides endpoints for the Taranis is an OSINT gathering and analysis tool.
  license:
    name: EUPL-1.2
  version: 0.1.8
...
servers:
  - url: /api
    description: default
  - url: https://dev.taranis.cyberrange.rocks/api
    description: dev
  - url: https://taranis.cyberrange.rocks/api
    description: main
...
security:
  - APIKey: []
  - UserAuth: []
paths:
...
  /config/reload-enum-dictionaries/{dictionary_type}:
    parameters:
      - name: dictionary_type
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [config]
      description: reload dictionaries
      # TODO: this should be an HTTP POST/PUT request
      responses:
        '200':
          description: successfully reloaded dictionaries
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/attributes:
    get:
      tags: [config]
      description: get attributes
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: attributes
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/attribute'
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    post:
      tags: [config]
      description: add attribute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/attribute'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/attributes/{attribute_id}:
    parameters:
      - name: attribute_id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [config]
      description: update attribute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/attribute'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    delete:
      tags: [config]
      description: delete attribute
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/attributes/{attribute_id}/enums:
    parameters:
      - name: attribute_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [config]
      description: get attribute enum
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: offset
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    post:
      tags: [config]
      description: add attribute enums
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                delete_existing:
                  type: boolean
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/attribute_enum'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/attributes/{attribute_id}/enums/{enum_id}:
    parameters:
      - name: attribute_id
        in: path
        required: true
        schema:
          type: integer
      - name: enum_id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [config]
      description: update attribute enum
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/attribute_enum'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    delete:
      tags: [config]
      description: delete attribute enum
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/report-item-types:
    get:
      tags: [config]
      description: get report item types
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/report_item_type.presentation'
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    post:
      tags: [config]
      description: add report item type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/report_item_type'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/report-item-types/{type_id}:
    parameters:
      - name: type_id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [config]
      description: update report item type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/report_item_type'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    delete:
      tags: [config]
      description: delete report item type
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/product-types:
    get:
      tags: [config]
      description: get all product types
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/product_type'
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    post:
      tags: [config]
      description: add product type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/product_type'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
...
  /config/permissions:
    get:
      tags: [config]
      description: get all permissions
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/permission'
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
...
components:
  schemas:
...
    notFound:
      description: "Not found response"
...
    resp.auth.error:
      type: object
      properties:
        error:
          type: string

    #general
...
    report_item_type:
      allOf:
        - type: object
          properties:
            id:
              type: integer
            title:
              type: string
            description:
              type: string
            attribute_groups:
              type: array
              items:
                $ref: '#/components/schemas/attribute_group'
    report_item_type.presentation:
      allOf:
        - $ref: '#/components/schemas/report_item_type'
        - $ref: '#/components/schemas/presentation'
...
    attribute_enum:
      type: object
      properties:
        id:
          type: integer
        index:
          type: integer
        value:
          type: string
        description:
          type: string
...
    attribute:
      allOf:
        - $ref: '#/components/schemas/attribute.base'
        - type: object
          properties:
            attribute_enums:
              type: array
              items:
                $ref: '#/components/schemas/attribute_enum'
...
  securitySchemes:
    APIKey:
      type: http
      scheme: bearer
      bearerFormat: APIKEY
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT